Rename package to azure-datalake-store
