+++ .coveragerc
@@ -1,9 +1,9 @@
 [run]
 include =
-    adlfs/*
+    azure/datalake/store/*
 
 omit =
-    adlfs/tests/test*
+    azure/datalake/store/tests/test*
 
 [report]
 show_missing = True
