+++ .travis.yml
@@ -17,7 +17,7 @@ install:
   - python setup.py develop
 
 script:
-  - py.test -x -vvv --doctest-modules --pyargs adlfs tests
+  - py.test -x -vvv --doctest-modules --pyargs azure.datalake.store tests
 
 notifications:
   email: false
