+++ MANIFEST.in
@@ -1,4 +1,4 @@
-recursive-include adlfs *.py
+recursive-include azure/datalake/store *.py
 recursive-include docs *.rst
 
 include setup.py
