+++ appveyor.yml
@@ -15,4 +15,4 @@ install:
 build: off
 
 test_script:
-  - "%PYTHON%\\Scripts\\pytest.exe -x -vvv --doctest-modules --pyargs adlfs tests"
+  - "%PYTHON%\\Scripts\\pytest.exe -x -vvv --doctest-modules --pyargs azure.datalake.store tests"
