+++ azure/datalake/store/core.py
