+++ azure/datalake/store/exceptions.py
