+++ azure/datalake/store/multithread.py
@@ -66,7 +66,7 @@ class ADLDownloader(object):
 
     See Also
     --------
-    adlfs.transfer.ADLTransferClient
+    azure.datalake.store.transfer.ADLTransferClient
     """
     def __init__(self, adlfs, rpath, lpath, nthreads=None, chunksize=2**28,
                  blocksize=2**22, client=None, run=True, overwrite=False):
@@ -231,7 +231,7 @@ class ADLUploader(object):
 
     See Also
     --------
-    adlfs.transfer.ADLTransferClient
+    azure.datalake.store.transfer.ADLTransferClient
     """
     def __init__(self, adlfs, rpath, lpath, nthreads=None, chunksize=2**28,
                  blocksize=2**25, client=None, run=True, delimiter=None,
