+++ setup.py
@@ -3,7 +3,7 @@
 import os
 from setuptools import setup
 
-setup(name='adlfs',
+setup(name='azure-datalake-store',
       version='0.0.1',
       description='Convenient Filesystem interface to Azure Data-lake Store',
       url='https://github.com/Azure/azure-data-lake-store-python',
@@ -11,7 +11,11 @@ setup(name='adlfs',
       maintainer_email='',
       license='',
       keywords='azure',
-      packages=['adlfs'],
+      packages=[
+          'azure',
+          'azure.datalake',
+          'azure.datalake.store'
+      ],
       install_requires=[open('requirements.txt').read().strip().split('\n')],
       long_description=(open('README.rst').read() if os.path.exists('README.rst')
                         else ''),
