+++ tests/README.md
@@ -1,6 +1,6 @@
 To run the test suite:
 
-    py.test -x -vvv --doctest-modules --pyargs adlfs tests
+    py.test -x -vvv --doctest-modules --pyargs azure-datalake-store tests
 
 This test suite uses [VCR.py](https://github.com/kevin1024/vcrpy) to record the
 responses from Azure. Borrowing from VCR's
@@ -11,11 +11,11 @@ invoking the test suite (defaults to `none`).
 
 To record responses for a new test without updating previous recordings:
 
-    RECORD_MODE=once py.test -x -vvv --doctest-modules --pyargs adlfs tests
+    RECORD_MODE=once py.test -x -vvv --doctest-modules --pyargs azure-datalake-store tests
 
 To record responses for all tests even if previous recordings exist:
 
-    RECORD_MODE=all py.test -x -vvv --doctest-modules --pyargs adlfs tests
+    RECORD_MODE=all py.test -x -vvv --doctest-modules --pyargs azure-datalake-store tests
 
 When recording new responses, you will need valid Azure credentials. The following
 environment variables should be defined:
